Corrección ejercicios 9 y 10 primera parte. Ejercicios 1 a 8 de la tercera parte resueltos. TPS 1, 2 y 3 resueltos.

+++ 1-EntradaSalida/jsEntradaSalida-10.js
@@ -5,7 +5,7 @@ mostrar el importe con un Descuento del 25 %
 en el cuadro de texto "RESULTADO"*/
 function mostrarAumento() {
 
-	let importe = parseInt(document.getElementById("txtIdImporte").value);
+	let importe = parseFloat(document.getElementById("txtIdImporte").value);
 	let nuevoImporte = importe - (importe * 0.25);
 	document.getElementById("txtIdResultado").value = nuevoImporte;
